fixed token & scripts

+++ stop.sh
@@ -1,3 +1,11 @@
 #!/bin/bash 
 
-pkill python3*
+
+if [[ -e process.id ]]; then
+  kill `cat process.id`
+  rm process.id
+  echo "Stopped"
+  else
+    pkill python3*
+    echo "Stopped forcefully"
+    fi
